markers now keep track of username and skill level
