+++ src/main/webapp/aboutus.html
@@ -35,7 +35,6 @@ limitations under the License.
     
   </head>
   <body onload="addLoginOrLogoutLinkToNavigation();">
-<<<<<<< HEAD
     <div class="container-fluid pt-5">	
         <nav class="navbar navbar-expand-sm navbar-light fixed-top justify-content-between" style="background-color: #FFA07A">
           <ul class="navbar-nav" id="navigation">
@@ -44,6 +43,8 @@ limitations under the License.
             <li class="nav-item"><a class="nav-link font" href="/feed.html">Message Feed</a></li>
             <li class="nav-item"><a class="nav-link font" href="/community.html">Community Page</a></li>
             <!-- <li class="nav-item"><a class="nav-link font" href="/sentiment.html">Sentiment</a></li>      -->
+            <li class="nav-item"><a class="nav-link" href="/map.html">Map</a></li>
+            <li class="nav-item"><a class="nav-link" href="/translate.html">Translate</a></li>
             <div class="divider"/>  
           </ul>
   
@@ -54,21 +55,6 @@ limitations under the License.
             </form>
           </div>
         </nav>
-=======
-    <div class="container-fluid">	
-      <nav class="navbar navbar-expand-sm bg-dark navbar-dark">
-        <ul class="navbar-nav" id="navigation">
-          <li class="nav-item"><a class="nav-link" href="/">Home</a></li>
-          <li class="nav-item active"><a class="nav-link" href="/aboutus.html">About Our Team</a></li>
-          <li class="nav-item"><a class="nav-link" href="/feed.html">Message Feed</a></li>
-          <li class="nav-item"><a class="nav-link" href="/community.html">Community Page</a></li>
-          <li class="nav-item"><a class="nav-link" href="/sentiment.html">Sentiment</a></li>
-          <li class="nav-item"><a class="nav-link" href="/translate.html">Translate</a></li>
-          <li class="nav-item"><a class="nav-link" href="/map.html">Map</a></li>
-        </ul>
-      </nav>
->>>>>>> 9d7a543082c563559e0b98a573a06e60fbb0bc5c
-
       <div class="pt-5">
         <h2 class="font-weight-light">About Our Team</h2>
