+++ src/main/webapp/community.html
@@ -51,7 +51,6 @@
   <body onload="buildUI(); addLoginOrLogoutLinkToNavigation();">
     <div class="container-fluid pt-5">	
 
-<<<<<<< HEAD
         <nav class="navbar navbar-expand-sm navbar-light fixed-top justify-content-between" style="background-color: #FFA07A">
           <ul class="navbar-nav" id="navigation">
             <li class="nav-item"><a class="nav-link" href="/">Home</a></li>
@@ -59,6 +58,8 @@
             <li class="nav-item"><a class="nav-link font" href="/feed.html">Message Feed</a></li>
             <li class="nav-item active"><a class="nav-link font" href="/community.html">Community Page</a></li>
             <!-- <li class="nav-item"><a class="nav-link font" href="/sentiment.html">Sentiment</a></li>      -->
+            <li class="nav-item"><a class="nav-link" href="/map.html">Map</a></li>
+            <li class="nav-item"><a class="nav-link" href="/translate.html">Translate</a></li>
             <div class="divider"/>  
           </ul>
   
@@ -69,19 +70,6 @@
             </form>
           </div>
         </nav>
-=======
-      <nav class="navbar navbar-expand-sm bg-dark navbar-dark">
-        <ul class="navbar-nav" id="navigation">
-          <li class="nav-item"><a class="nav-link" href="/">Home</a></li>
-          <li class="nav-item"><a class="nav-link" href="/aboutus.html">About Our Team</a></li>
-          <li class="nav-item"><a class="nav-link" href="/feed.html">Message Feed</a></li>
-          <li class="nav-item active"><a class="nav-link" href="/community.html">Community Page</a></li>
-          <li class="nav-item"><a class="nav-link" href="/sentiment.html">Sentiment</a></li>
-          <li class="nav-item"><a class="nav-link" href="/translate.html">Translate</a></li>
-          <li class="nav-item"><a class="nav-link" href="/map.html">Map</a></li>
-        </ul>
-      </nav>
->>>>>>> 9d7a543082c563559e0b98a573a06e60fbb0bc5c
 
       <div class="pt-5" id="content">
         <h1>Community Page</h1>
