+++ src/main/webapp/feed.html
@@ -25,15 +25,9 @@
         <li class="nav-item active"><a class="nav-link font" href="/feed.html">Message Feed</a></li>
         <li class="nav-item"><a class="nav-link font" href="/community.html">Community Page</a></li>
         <!-- <li class="nav-item"><a class="nav-link font" href="/sentiment.html">Sentiment</a></li>      -->
-        <div class="divider"/>  
-=======
-        <li class="nav-item"><a class="nav-link" href="/aboutus.html">About Our Team</a></li>
-        <li class="nav-item active"><a class="nav-link" href="/feed.html">Message Feed</a></li>
-        <li class="nav-item"><a class="nav-link" href="/community.html">Community Page</a></li>
-        <li class="nav-item"><a class="nav-link" href="/sentiment.html">Sentiment</a></li>
-        <li class="nav-item"><a class="nav-link" href="/translate.html">Translate</a></li>
         <li class="nav-item"><a class="nav-link" href="/map.html">Map</a></li>
->>>>>>> 9d7a543082c563559e0b98a573a06e60fbb0bc5c
+        <li class="nav-item"><a class="nav-link" href="/translate.html">Translate</a></li>
+        <div class="divider"/>  
       </ul>
 
       <div class="search-container">
