+++ src/main/webapp/index.html
@@ -38,20 +38,13 @@ limitations under the License.
      <nav class="navbar navbar-expand-sm navbar-light fixed-top justify-content-between" style="background-color: #FFA07A">
         <ul class="navbar-nav" id="navigation">
           <li class="nav-item active"><a class="nav-link" href="/">Home</a></li>
-<<<<<<< HEAD
           <li class="nav-item"><a class="nav-link font" href="/aboutus.html">About Our Team</a></li>
           <li class="nav-item"><a class="nav-link font" href="/feed.html">Message Feed</a></li>
           <li class="nav-item"><a class="nav-link font" href="/community.html">Community Page</a></li>
           <!-- <li class="nav-item"><a class="nav-link font" href="/sentiment.html">Sentiment</a></li>      -->
-          <div class="divider"/>  
-=======
-          <li class="nav-item"><a class="nav-link" href="/aboutus.html">About Our Team</a></li>
-          <li class="nav-item"><a class="nav-link" href="/feed.html">Message Feed</a></li>
-          <li class="nav-item"><a class="nav-link" href="/community.html">Community Page</a></li>
-          <li class="nav-item"><a class="nav-link" href="/sentiment.html">Sentiment</a></li>
+          <li class="nav-item"><a class="nav-link" href="/map.html">Map</a></li>
           <li class="nav-item"><a class="nav-link" href="/translate.html">Translate</a></li>
-          <li class="nav-item"><a class="nav-link" href="/map.html">Map</a></li>          
->>>>>>> 9d7a543082c563559e0b98a573a06e60fbb0bc5c
+          <div class="divider"/>  
         </ul>
 
         <div class="search-container">
