+++ src/main/webapp/map.html
@@ -2,32 +2,46 @@
 <html>
 <head>
   <title>Marker Storage</title>
-<<<<<<< HEAD
+  <link rel="stylesheet" href="/css/main.css">
+    <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.5.0/css/all.css" integrity="sha384-B4dIYHKNBt8Bc12p+WXckhzcICo0wtJAoU8YZTY5qE0Id1GSseTk6S+L3BlXeVIU" crossorigin="anonymous">
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js" integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q" crossorigin="anonymous"></script>
+    <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js" integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl" crossorigin="anonymous"></script>
+    <script src="/js/navigation-loader.js"></script>
+    <script src="/js/search-results-loader.js"></script>
   <style>
     #map{
       width: 500px;
       height: 500px;
     }
   </style>
- </head>
- <body>
+  </head>
+  <body onload="addLoginOrLogoutLinkToNavigation();">
+  <div class="container-fluid pt-5">	
+
+        <nav class="navbar navbar-expand-sm navbar-light fixed-top justify-content-between" style="background-color: #FFA07A">
+          <ul class="navbar-nav" id="navigation">
+            <li class="nav-item"><a class="nav-link" href="/">Home</a></li>
+            <li class="nav-item"><a class="nav-link font" href="/aboutus.html">About Our Team</a></li>
+            <li class="nav-item"><a class="nav-link font" href="/feed.html">Message Feed</a></li>
+            <li class="nav-item"><a class="nav-link font" href="/community.html">Community Page</a></li>
+            <!-- <li class="nav-item"><a class="nav-link font" href="/sentiment.html">Sentiment</a></li>      -->
+            <li class="nav-item active"><a class="nav-link" href="/map.html">Map</a></li>
+            <li class="nav-item"><a class="nav-link" href="/translate.html">Translate</a></li>
+            <div class="divider"/>  
+          </ul>
+  
+          <div class="search-container">
+            <form class="form-inline" action="javascript:;" onsubmit="sendInput()" id="search-form">
+              <input class="mr-sm-2" type="text" placeholder="Search Buddies" id="search-input">
+              <button class="my-2 my-sm-0" type="submit"><i class="fa fa-search"></i></button>
+            </form>
+          </div>
+        </nav>
   <h1>Marker Storage</h1>
   <div id="map"></div>
   <p>Click in the map to create a new marker.</p>
-=======
-  
-  <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyAKK0F_x99qM2kfXHn9hNMxUqenT8CNa-E"></script>
-    <link rel="stylesheet" href="/css/main.css">
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js" integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q" crossorigin="anonymous"></script>
-    <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js" integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl" crossorigin="anonymous"></script>
-
-    <script src="/js/navigation-loader.js"></script>
-</head>
-
-<script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyAKK0F_x99qM2kfXHn9hNMxUqenT8CNa-E"></script>
->>>>>>> 9d7a543082c563559e0b98a573a06e60fbb0bc5c
   <script>
     let map;
     /* Editable marker that displays when a user clicks in the map. */
@@ -49,39 +63,36 @@
         return response.json();
       }).then((markers) => {
         markers.forEach((marker) => {
-         createMarkerForDisplay(marker.lat, marker.lng, marker.content, marker.userName, marker.skillLevel)
+         createMarkerForDisplay(marker.lat, marker.lng, marker.content)
         });
       });
     }
     /** Creates a marker that shows a read-only info window when clicked. */
-    function createMarkerForDisplay(lat, lng, content, userName, skillLevel){
-    	let url = "http://maps.google.com/mapfiles/ms/icons/green-dot.png";
+    function createMarkerForDisplay(lat, lng, content){
+    	  let url = "http://maps.google.com/mapfiles/ms/icons/green-dot.png";
       const marker = new google.maps.Marker({
         position: {lat: lat, lng: lng},
         map: map,
         icon: url
       });
       var infoWindow = new google.maps.InfoWindow({
-        content: content
+    	  content: content
       });
       marker.addListener('click', () => {
         infoWindow.open(map, marker);
       });
     }
     /** Sends a marker to the backend for saving. */
-    function postMarker(lat, lng, content, userName, skillLevel){
+    function postMarker(lat, lng, content){
       const params = new URLSearchParams();
       params.append('lat', lat);
       params.append('lng', lng);
       params.append('content', content);
-      params.append('userName', userName);
-      params.append('skillLevel', skillLevel);
       fetch('/markers', {
         method: 'POST',
         body: params
       });
-    } 
-    
+    }
     /** Creates a marker that shows a textbox the user can edit. */
     function createMarkerForEdit(lat, lng){
       // If we're already showing an editable marker, then remove it.
@@ -118,31 +129,8 @@
       return containerDiv;
     }
   </script>
-</head>
-<<<<<<< HEAD
   <script async defer
     src="https://maps.googleapis.com/maps/api/js?key=AIzaSyAKK0F_x99qM2kfXHn9hNMxUqenT8CNa-E&callback=createMap">
   </script>
-=======
-
-<body onload="addLoginOrLogoutLinkToNavigation(); createMap();">
-  <div class="container-fluid">	
-  
-    <nav class="navbar navbar-expand-sm bg-dark navbar-dark">
-      <ul class="navbar-nav" id="navigation">
-        <li class="nav-item"><a class="nav-link" href="/">Home</a></li>
-        <li class="nav-item"><a class="nav-link" href="/aboutus.html">About Our Team</a></li>
-        <li class="nav-item"><a class="nav-link" href="/feed.html">Message Feed</a></li>
-        <li class="nav-item"><a class="nav-link" href="/community.html">Community Page</a></li>
-        <li class="nav-item"><a class="nav-link" href="/sentiment.html">Sentiment</a></li>
-        <li class="nav-item"><a class="nav-link" href="/translate.html">Translate</a></li>
-        <li class="nav-item active"><a class="nav-link" href="/map.html">Map</a></li>   
-      </ul>
-    </nav>
-  
-  <h1>Marker Storage</h1>
-  <div id="map"></div>
-  <p>Click in the map to create a new marker.</p>
->>>>>>> 9d7a543082c563559e0b98a573a06e60fbb0bc5c
 </body>
 </html>
