+++ src/main/webapp/search-results.html
@@ -21,7 +21,9 @@
                     <li class="nav-item"><a class="nav-link font" href="/aboutus.html">About Our Team</a></li>
                     <li class="nav-item"><a class="nav-link font" href="/feed.html">Message Feed</a></li>
                     <li class="nav-item"><a class="nav-link font" href="/community.html">Community Page</a></li>
-                    <li class="nav-item"><a class="nav-link font" href="/sentiment.html">Sentiment</a></li>     
+                    <li class="nav-item"><a class="nav-link font" href="/sentiment.html">Sentiment</a></li> 
+                    <li class="nav-item"><a class="nav-link" href="/map.html">Map</a></li>
+                    <li class="nav-item"><a class="nav-link" href="/translate.html">Translate</a></li>    
                     <div class="divider"/>  
                 </ul>
