+++ src/main/webapp/translate.html
@@ -23,8 +23,8 @@
         <li class="nav-item"><a class="nav-link" href="/feed.html">Message Feed</a></li>
         <li class="nav-item"><a class="nav-link" href="/community.html">Community Page</a></li>
         <li class="nav-item"><a class="nav-link" href="/sentiment.html">Sentiment</a></li>
-        <li class="nav-item active"><a class="nav-link" href="/translate.html">Translate</a></li>
-        <li class="nav-item"><a class="nav-link" href="/map.html">Map</a></li>   
+        <li class="nav-item"><a class="nav-link" href="/map.html">Map</a></li>
+        <li class="nav-item active"><a class="nav-link" href="/translate.html">Translate</a></li>  
       </ul>
     </nav>
 
@@ -71,11 +71,7 @@
     <textarea id="text"></textarea>
 
     <select id="language">
-<<<<<<< HEAD
       <option value="en">English</option>
-=======
-    	  <option value="en">English</option>
->>>>>>> 9d7a543082c563559e0b98a573a06e60fbb0bc5c
       <option value="zh">Chinese</option>
       <option value="es">Spanish</option>
       <option value="hi">Hindi</option>
